+++ app/about/page.tsx
@@ -1,7 +1,6 @@
 "use client";
 
 import SquareArrowLeft from "lucide-react/dist/esm/icons/square-arrow-left";
-import "./Page.styled.tsx";
 import {
   About,
   AboutSection,
