+++ app/styles/styled.d.ts
@@ -1,5 +1,5 @@
 import "styled-components";
-import type { ThemeType } from "./themes";
+import type { ThemeType } from "../types/themes";
 
 declare module "styled-components" {
   export interface DefaultTheme extends ThemeType {}
