--- app/styles/themes.ts
@@ -1,240 +1,0 @@
-export type ThemeType = {
-  name: string;
-  background: string;
-  backgroundSecondary: string;
-  backgroundGradient: string;
-  backgroundGradientTimer: string;
-  text: string;
-  buttonText: string;
-  buttonTextDisabled: string;
-  buttonBackgroundActive: string;
-  inputInner: string;
-  highlight: string;
-  boxShadowInsetSoft: string;
-  boxShadowOuter: string;
-  boxShadowOuterStrong: string;
-  boxShadowTimer: string;
-  boxShadowTimerStrong: string;
-  boxShadowInputDisabled: string;
-  boxShadowInput: string;
-};
-
-export const primary: ThemeType = {
-  name: "primary",
-  background: "#e1dfdf",
-  backgroundSecondary: "#edececaa",
-  backgroundGradient: "linear-gradient(145deg, #f8f8f8, #ededed)",
-  backgroundGradientTimer: "linear-gradient(145deg, #f1efef, #cbc9c9)",
-  text: "#333333",
-  buttonText: "#252525",
-  buttonTextDisabled: "#878787f2",
-  buttonBackgroundActive: "#e1dfdf",
-  inputInner: "#eae7e7f2",
-  highlight: "#b62c0a",
-  boxShadowInsetSoft: "inset 6px 6px 13px #c8c6c6, inset -6px -6px 13px #faf8f8",
-  boxShadowOuter: "10px 10px 20px #989797, -10px -10px 22px #eeeded",
-  boxShadowOuterStrong: "6px 6px 22px #989797, -6px -6px 22px #eeeded",
-  boxShadowTimer: "inset 5px 5px 10px #bebebe, inset -5px -5px 10px #ffffff",
-  boxShadowTimerStrong: "15px 15px 30px #b6b6b6, -5px -5px 30px #e4e4e4",
-  boxShadowInputDisabled: "inset 1px 1px 5px #bebebe, inset -5px -5px 10px #eae6e6ff",
-  boxShadowInput: "inset 5px 5px 7px #c7c7c7, inset -5px -5px 7px #e0dadaff",
-};
-
-export const calm: ThemeType = {
-  name: "calm",
-  background: "#dde6f5",
-  backgroundSecondary: "#e6eef9cc",
-  backgroundGradient: "linear-gradient(145deg, #eaf1fb, #cfdff1)",
-  backgroundGradientTimer: "linear-gradient(145deg, #dce7f3, #b7d0e7ff)",
-  text: "#1a1a1a",
-  buttonText: "#1f2d3d",
-  buttonTextDisabled: "#7287a0f2",
-  buttonBackgroundActive: "#dde6f5",
-  inputInner: "#edf3fbf2",
-  highlight: "#1e63b1",
-  boxShadowInsetSoft: "inset 6px 6px 13px #c2cfe0, inset -6px -6px 13px #e8f1faff",
-  boxShadowOuter: "10px 10px 20px #a2b1c3, -10px -10px 22px #e8f1faff",
-  boxShadowOuterStrong: "6px 6px 22px #a2b1c3, -6px -6px 10px #d5dfeaff",
-  boxShadowTimer: "inset 5px 5px 10px #c9d7e8, inset -5px -5px 10px #e8f1faff",
-  boxShadowTimerStrong: "15px 15px 30px #b0c2d8, -5px -5px 30px #e3edf5",
-  boxShadowInputDisabled: "inset 1px 1px 5px #c9d7e8, inset -5px -5px 10px #e8f1faff",
-  boxShadowInput: "inset 5px 5px 7px #ccd7e3, inset -5px -5px 7px #e8eff5",
-};
-
-export const peach: ThemeType = {
-  name: "peach",
-  background: "#f9ece6",
-  backgroundSecondary: "#faefebcc",
-  backgroundGradient: "linear-gradient(145deg, #fef5f1, #f1dcd4)",
-  backgroundGradientTimer: "linear-gradient(145deg, #f7e8e1, #e5c9be)",
-  text: "#2d1c17",
-  buttonText: "#3e2b26",
-  buttonTextDisabled: "#9e7e75f2",
-  buttonBackgroundActive: "#f9ece6",
-  inputInner: "#fdf5f1f2",
-  highlight: "#d17659",
-  boxShadowInsetSoft: "inset 6px 6px 13px #e5d6d0, inset -6px -6px 13px #fff9f6",
-  boxShadowOuter: "10px 10px 20px #d4c2bb, -10px -10px 22px #fff6f2",
-  boxShadowOuterStrong: "6px 6px 22px #d4c2bb, -6px -6px 22px #fff6f2",
-  boxShadowTimer: "inset 5px 5px 10px #decac3, inset -5px -5px 10px #f2e4df",
-  boxShadowTimerStrong: "15px 15px 30px #cbaea7, -5px -5px 30px #f2e4df",
-  boxShadowInputDisabled: "inset 1px 1px 5px #decac3, inset -5px -5px 10px #f2e4df",
-  boxShadowInput: "inset 5px 5px 7px #e1d0cb, inset -5px -5px 7px #f7efeb",
-};
-export const lilac: ThemeType = {
-  name: "lilac",
-  background: "#f1e7f7",
-  backgroundSecondary: "#f5eefacc",
-  backgroundGradient: "linear-gradient(145deg, #fdf7ff, #e1d5ec)",
-  backgroundGradientTimer: "linear-gradient(145deg, #eee1f6, #ceb9da)",
-  text: "#2d2536",
-  buttonText: "#4b3a61",
-  buttonTextDisabled: "#a58cb1f2",
-  buttonBackgroundActive: "#f1e7f7",
-  inputInner: "#f6effaf2",
-  highlight: "#9646c0",
-  boxShadowInsetSoft: "inset 6px 6px 13px #d8c9e4, inset -6px -6px 13px #e1d5ebff",
-  boxShadowOuter: "10px 10px 20px #c4b3d1, -10px -10px 22px #f5ecf9",
-  boxShadowOuterStrong: "6px 6px 22px #c4b3d1, -6px -6px 22px #f5ecf9",
-  boxShadowTimer: "inset 5px 5px 10px #d4c1e0, inset -5px -5px 10px #e1d5ebff",
-  boxShadowTimerStrong: "15px 15px 30px #baa4c9, -5px -5px 30px #f5ecf9",
-  boxShadowInputDisabled: "inset 1px 1px 5px #d4c1e0, inset -5px -5px 10px #e1d5ebff",
-  boxShadowInput: "inset 5px 5px 7px #e2d3ee, inset -5px -5px 7px #e1d5ebff",
-};
-export const sunny: ThemeType = {
-  name: "sunny",
-  background: "#fff4e5",
-  backgroundSecondary: "#fef2e0cc",
-  backgroundGradient: "linear-gradient(145deg, #fffaf3, #fbe5cf)",
-  backgroundGradientTimer: "linear-gradient(145deg, #fce7d5, #ecdfd4ff)",
-  text: "#3e2c1e",
-  buttonText: "#4a3625",
-  buttonTextDisabled: "#a58a74f2",
-  buttonBackgroundActive: "#fff4e5",
-  inputInner: "#fff6ebf2",
-  highlight: "#ff8f3c",
-  boxShadowInsetSoft: "inset 6px 6px 13px #e0cbb8, inset -6px -6px 13px #d6d6d6ff",
-  boxShadowOuter: "10px 10px 20px #d4c3b2, -10px -10px 22px #fffaf4",
-  boxShadowOuterStrong: "6px 6px 22px #d4c3b2, -6px -6px 22px #fffaf4",
-  boxShadowTimer: "inset 5px 5px 10px #e2cfbc, inset -5px -5px 10px #d6d6d6ff",
-  boxShadowTimerStrong: "15px 15px 30px #c9b29a, -5px -5px 30px #fef0e3",
-  boxShadowInputDisabled: "inset 1px 1px 5px #e2cfbc, inset -5px -5px 10px #d6d6d6ff",
-  boxShadowInput: "inset 5px 5px 7px #eedcca, inset -5px -5px 7px #d6d6d6ff",
-};
-export const lime: ThemeType = {
-  name: "lime",
-  background: "#eefae1",
-  backgroundSecondary: "#f7fce7",
-  backgroundGradient: "linear-gradient(145deg, #f4ffe5, #ddf0c7)",
-  backgroundGradientTimer: "linear-gradient(145deg, #e6f7d1, #daeacaff)",
-  text: "#3a4b1d",
-  buttonText: "#2a3b11",
-  buttonTextDisabled: "#7d8f66",
-  buttonBackgroundActive: "#eefae1",
-  inputInner: "#f2fce7",
-  highlight: "#96b40f",
-  boxShadowInsetSoft: "inset 6px 6px 13px #cdd8bc, inset -6px -6px 13px #eef5e1ff",
-  boxShadowOuter: "10px 10px 20px #b5c3a1, -10px -10px 22px #f2fbe4",
-  boxShadowOuterStrong: "6px 6px 22px #a9b68f, -6px -6px 22px #e8f9d6",
-  boxShadowTimer: "inset 5px 5px 10px #c1d3a9, inset -5px -5px 10px #fafff1",
-  boxShadowTimerStrong: "15px 15px 30px #b3cba2, -5px -5px 30px #e1f5c4",
-  boxShadowInputDisabled: "inset 1px 1px 5px #c4d7a4, inset -5px -5px 10px #eef5e1ff",
-  boxShadowInput: "inset 5px 5px 7px #d2e5b8, inset -5px -5px 7px #eef5e1ff",
-};
-export const breeze: ThemeType = {
-  name: "breeze",
-  background: "#e6fff8",
-  backgroundSecondary: "#d9fdf3cc",
-  backgroundGradient: "linear-gradient(145deg, #f0fffc, #c5f2e7)",
-  backgroundGradientTimer: "linear-gradient(145deg, #d7f8ef, #b8e8dc)",
-  text: "#234941",
-  buttonText: "#1f3d39",
-  buttonTextDisabled: "#7da89df2",
-  buttonBackgroundActive: "#e6fff8",
-  inputInner: "#edfffaf2",
-  highlight: "#28c7a5",
-  boxShadowInsetSoft: "inset 6px 6px 13px #bfe4d8, inset -6px -6px 13px #d2ede4ff",
-  boxShadowOuter: "10px 10px 20px #b2d8cf, -10px -10px 22px #d2ede4ff",
-  boxShadowOuterStrong: "6px 6px 22px #b2d8cf, -6px -6px 22px #d2ede4ff",
-  boxShadowTimer: "inset 5px 5px 10px #c2e7dc, inset -5px -5px 10px #d2ede4ff",
-  boxShadowTimerStrong: "15px 15px 30px #a3ccc0, -5px -5px 30px #eafff8",
-  boxShadowInputDisabled: "inset 1px 1px 5px #c2e7dc, inset -5px -5px 10px #d2ede4ff",
-  boxShadowInput: "inset 5px 5px 7px #d6f6ec, inset -5px -5px 7px #dcf1eaff",
-};
-export const ocean: ThemeType = {
-  name: "ocean",
-  background: "#dceff5",
-  backgroundSecondary: "#e8f6fa",
-  backgroundGradient: "linear-gradient(145deg, #eaf9fc, #c4e3ea)",
-  backgroundGradientTimer: "linear-gradient(145deg, #d5edf5, #a5cfd9)",
-  text: "#1d3d4f",
-  buttonText: "#13303f",
-  buttonTextDisabled: "#7a9da9",
-  buttonBackgroundActive: "#dceff5",
-  inputInner: "#e1f2f8",
-  highlight: "#007a99",
-  boxShadowInsetSoft: "inset 6px 6px 13px #bcd3d9, inset -6px -6px 13px #e7f5f7ff",
-  boxShadowOuter: "10px 10px 20px #a1b5bb, -10px -10px 22px #f0fdff",
-  boxShadowOuterStrong: "6px 6px 22px #99b1b7, -6px -6px 22px #e1f7fc",
-  boxShadowTimer: "inset 5px 5px 10px #b0cfd6, inset -5px -5px 10px #e7f5f7ff",
-  boxShadowTimerStrong: "15px 15px 30px #a2c2c9, -5px -5px 30px #d5f5fb",
-  boxShadowInputDisabled: "inset 1px 1px 5px #a2bcc1, inset -5px -5px 10px #e7f5f7ff",
-  boxShadowInput: "inset 5px 5px 7px #bedbe2, inset -5px -5px 7px #e7f5f7ff",
-};
-export const sand: ThemeType = {
-  name: "sand",
-  background: "#f3ece6",
-  backgroundSecondary: "#f9f4f0",
-  backgroundGradient: "linear-gradient(145deg, #f7f1ed, #e9ded7)",
-  backgroundGradientTimer: "linear-gradient(145deg, #f4e7df, #d9c7bd)",
-  text: "#4d2c23",
-  buttonText: "#3e241d",
-  buttonTextDisabled: "#a18a82",
-  buttonBackgroundActive: "#f3ece6",
-  inputInner: "#f7f0eb",
-
-  highlight: "#bf5a3c",
-
-  boxShadowInsetSoft: "inset 6px 6px 13px #d6c8c1, inset -6px -6px 13px #fff9f4",
-  boxShadowOuter: "10px 10px 20px #b9aca6, -10px -10px 22px #fff6f1",
-  boxShadowOuterStrong: "6px 6px 22px #b09f97, -6px -6px 22px #f8eee9",
-  boxShadowTimer: "inset 5px 5px 10px #cab8b0, inset -5px -5px 10px #fffaf6",
-  boxShadowTimerStrong: "15px 15px 30px #c0aca5, -5px -5px 30px #f2e4dc",
-  boxShadowInputDisabled: "inset 1px 1px 5px #d4c4bc, inset -5px -5px 10px #fff9f5",
-  boxShadowInput: "inset 5px 5px 7px #e1d3cc, inset -5px -5px 7px #f8eee9",
-};
-export const moss: ThemeType = {
-  name: "moss",
-  background: "#e8ebe4",
-  backgroundSecondary: "#f2f4ef",
-  backgroundGradient: "linear-gradient(145deg, #f1f3ed, #dfe3db)",
-  backgroundGradientTimer: "linear-gradient(145deg, #e1e5dd, #c5c8bf)",
-  text: "#2f372f",
-  buttonText: "#263026",
-  buttonTextDisabled: "#7f8c7f",
-  buttonBackgroundActive: "#e8ebe4",
-  inputInner: "#f4f6f0",
-  highlight: "#5b7b2d",
-  boxShadowInsetSoft: "inset 6px 6px 13px #cdd0c9, inset -6px -6px 13px #eef3e2ff",
-  boxShadowOuter: "10px 10px 20px #b5b8b1, -10px -10px 22px #eef1e6ff",
-  boxShadowOuterStrong: "6px 6px 22px #a6aaa3, -6px -6px 22px #eceeea",
-  boxShadowTimer: "inset 5px 5px 10px #c1c5be, inset -5px -5px 10px #eef3e2ff",
-  boxShadowTimerStrong: "15px 15px 30px #b4b7b1, -5px -5px 30px #ebeee8",
-  boxShadowInputDisabled: "inset 1px 1px 5px #c7cbc4, inset -5px -5px 10px #fafcf8",
-  boxShadowInput: "inset 5px 5px 7px #d5d8d2, inset -5px -5px 7px #eef3e2ff",
-};
-
-export const themeMeta = [
-  { name: "primary", highlight: "#b62c0a", backgroundGradientTimer: "linear-gradient(145deg, #f1efef, #cbc9c9)" },
-  { name: "calm", highlight: "#1e63b1", backgroundGradientTimer: "linear-gradient(145deg, #dce7f3, #b7d0e7ff)" },
-  { name: "peach", highlight: "#d17659", backgroundGradientTimer: "linear-gradient(145deg, #f7e8e1, #e5c9be)" },
-  { name: "lilac", highlight: "#9646c0", backgroundGradientTimer: "linear-gradient(145deg, #eee1f6, #ceb9da)" },
-  { name: "sunny", highlight: "#ff8f3c", backgroundGradientTimer: "linear-gradient(145deg, #fce7d5, #ecdfd4ff)" },
-  { name: "lime", highlight: "#96b40f", backgroundGradientTimer: "linear-gradient(145deg, #e6f7d1, #daeacaff)" },
-  { name: "breeze", highlight: "#28c7a5", backgroundGradientTimer: "linear-gradient(145deg, #d7f8ef, #b8e8dc)" },
-  { name: "ocean", highlight: "#007a99", backgroundGradientTimer: "linear-gradient(145deg, #d5edf5, #a5cfd9)" },
-  { name: "sand", highlight: "#bf5a3c", backgroundGradientTimer: "linear-gradient(145deg, #f4e7df, #d9c7bd)" },
-  { name: "moss", highlight: "#5b7b2d", backgroundGradientTimer: "linear-gradient(145deg, #e1e5dd, #c5c8bf)" },
-] as const;
-
-export type ThemeName = (typeof themeMeta)[number]["name"];
