+++ app/styles/themes/breeze.ts
@@ -1,0 +1,22 @@
+import { ThemeType } from "../../types/themes";
+
+export const breeze: ThemeType = {
+  name: "breeze",
+  background: "#e6fff8",
+  backgroundSecondary: "#d9fdf3cc",
+  backgroundGradient: "linear-gradient(145deg, #f0fffc, #c5f2e7)",
+  backgroundGradientTimer: "linear-gradient(145deg, #d7f8ef, #b8e8dc)",
+  text: "#234941",
+  buttonText: "#1f3d39",
+  buttonTextDisabled: "#7da89df2",
+  buttonBackgroundActive: "#e6fff8",
+  inputInner: "#edfffaf2",
+  highlight: "#28c7a5",
+  boxShadowInsetSoft: "inset 6px 6px 13px #bfe4d8, inset -6px -6px 13px #d2ede4ff",
+  boxShadowOuter: "10px 10px 20px #b2d8cf, -10px -10px 22px #d2ede4ff",
+  boxShadowOuterStrong: "6px 6px 22px #b2d8cf, -6px -6px 22px #d2ede4ff",
+  boxShadowTimer: "inset 5px 5px 10px #c2e7dc, inset -5px -5px 10px #d2ede4ff",
+  boxShadowTimerStrong: "15px 15px 30px #a3ccc0, -5px -5px 30px #eafff8",
+  boxShadowInputDisabled: "inset 1px 1px 5px #c2e7dc, inset -5px -5px 10px #d2ede4ff",
+  boxShadowInput: "inset 5px 5px 7px #d6f6ec, inset -5px -5px 7px #dcf1eaff",
+};
