+++ app/styles/themes/calm.ts
@@ -1,0 +1,22 @@
+import { ThemeType } from "../../types/themes";
+
+export const calm: ThemeType = {
+  name: "calm",
+  background: "#dde6f5",
+  backgroundSecondary: "#e6eef9cc",
+  backgroundGradient: "linear-gradient(145deg, #eaf1fb, #cfdff1)",
+  backgroundGradientTimer: "linear-gradient(145deg, #dce7f3, #b7d0e7ff)",
+  text: "#1a1a1a",
+  buttonText: "#1f2d3d",
+  buttonTextDisabled: "#7287a0f2",
+  buttonBackgroundActive: "#dde6f5",
+  inputInner: "#edf3fbf2",
+  highlight: "#1e63b1",
+  boxShadowInsetSoft: "inset 6px 6px 13px #c2cfe0, inset -6px -6px 13px #e8f1faff",
+  boxShadowOuter: "10px 10px 20px #a2b1c3, -10px -10px 22px #e8f1faff",
+  boxShadowOuterStrong: "6px 6px 22px #a2b1c3, -6px -6px 10px #d5dfeaff",
+  boxShadowTimer: "inset 5px 5px 10px #c9d7e8, inset -5px -5px 10px #e8f1faff",
+  boxShadowTimerStrong: "15px 15px 30px #b0c2d8, -5px -5px 30px #e3edf5",
+  boxShadowInputDisabled: "inset 1px 1px 5px #c9d7e8, inset -5px -5px 10px #e8f1faff",
+  boxShadowInput: "inset 5px 5px 7px #ccd7e3, inset -5px -5px 7px #e8eff5",
+};
