+++ app/styles/themes/lilac.ts
@@ -1,0 +1,22 @@
+import { ThemeType } from "../../types/themes";
+
+export const lilac: ThemeType = {
+  name: "lilac",
+  background: "#f1e7f7",
+  backgroundSecondary: "#f5eefacc",
+  backgroundGradient: "linear-gradient(145deg, #fdf7ff, #e1d5ec)",
+  backgroundGradientTimer: "linear-gradient(145deg, #eee1f6, #ceb9da)",
+  text: "#2d2536",
+  buttonText: "#4b3a61",
+  buttonTextDisabled: "#a58cb1f2",
+  buttonBackgroundActive: "#f1e7f7",
+  inputInner: "#f6effaf2",
+  highlight: "#9646c0",
+  boxShadowInsetSoft: "inset 6px 6px 13px #d8c9e4, inset -6px -6px 13px #e1d5ebff",
+  boxShadowOuter: "10px 10px 20px #c4b3d1, -10px -10px 22px #f5ecf9",
+  boxShadowOuterStrong: "6px 6px 22px #c4b3d1, -6px -6px 22px #f5ecf9",
+  boxShadowTimer: "inset 5px 5px 10px #d4c1e0, inset -5px -5px 10px #e1d5ebff",
+  boxShadowTimerStrong: "15px 15px 30px #baa4c9, -5px -5px 30px #f5ecf9",
+  boxShadowInputDisabled: "inset 1px 1px 5px #d4c1e0, inset -5px -5px 10px #e1d5ebff",
+  boxShadowInput: "inset 5px 5px 7px #e2d3ee, inset -5px -5px 7px #e1d5ebff",
+};
