+++ app/styles/themes/lime.ts
@@ -1,0 +1,22 @@
+import { ThemeType } from "../../types/themes";
+
+export const lime: ThemeType = {
+  name: "lime",
+  background: "#eefae1",
+  backgroundSecondary: "#f7fce7",
+  backgroundGradient: "linear-gradient(145deg, #f4ffe5, #ddf0c7)",
+  backgroundGradientTimer: "linear-gradient(145deg, #e6f7d1, #daeacaff)",
+  text: "#3a4b1d",
+  buttonText: "#2a3b11",
+  buttonTextDisabled: "#7d8f66",
+  buttonBackgroundActive: "#eefae1",
+  inputInner: "#f2fce7",
+  highlight: "#96b40f",
+  boxShadowInsetSoft: "inset 6px 6px 13px #cdd8bc, inset -6px -6px 13px #eef5e1ff",
+  boxShadowOuter: "10px 10px 20px #b5c3a1, -10px -10px 22px #f2fbe4",
+  boxShadowOuterStrong: "6px 6px 22px #a9b68f, -6px -6px 22px #e8f9d6",
+  boxShadowTimer: "inset 5px 5px 10px #c1d3a9, inset -5px -5px 10px #fafff1",
+  boxShadowTimerStrong: "15px 15px 30px #b3cba2, -5px -5px 30px #e1f5c4",
+  boxShadowInputDisabled: "inset 1px 1px 5px #c4d7a4, inset -5px -5px 10px #eef5e1ff",
+  boxShadowInput: "inset 5px 5px 7px #d2e5b8, inset -5px -5px 7px #eef5e1ff",
+};
