+++ app/styles/themes/moss.ts
@@ -1,0 +1,22 @@
+import { ThemeType } from "../../types/themes";
+
+export const moss: ThemeType = {
+  name: "moss",
+  background: "#e8ebe4",
+  backgroundSecondary: "#f2f4ef",
+  backgroundGradient: "linear-gradient(145deg, #f1f3ed, #dfe3db)",
+  backgroundGradientTimer: "linear-gradient(145deg, #e1e5dd, #c5c8bf)",
+  text: "#2f372f",
+  buttonText: "#263026",
+  buttonTextDisabled: "#7f8c7f",
+  buttonBackgroundActive: "#e8ebe4",
+  inputInner: "#f4f6f0",
+  highlight: "#5b7b2d",
+  boxShadowInsetSoft: "inset 6px 6px 13px #cdd0c9, inset -6px -6px 13px #eef3e2ff",
+  boxShadowOuter: "10px 10px 20px #b5b8b1, -10px -10px 22px #eef1e6ff",
+  boxShadowOuterStrong: "6px 6px 22px #a6aaa3, -6px -6px 22px #eceeea",
+  boxShadowTimer: "inset 5px 5px 10px #c1c5be, inset -5px -5px 10px #eef3e2ff",
+  boxShadowTimerStrong: "15px 15px 30px #b4b7b1, -5px -5px 30px #ebeee8",
+  boxShadowInputDisabled: "inset 1px 1px 5px #c7cbc4, inset -5px -5px 10px #fafcf8",
+  boxShadowInput: "inset 5px 5px 7px #d5d8d2, inset -5px -5px 7px #eef3e2ff",
+};
