+++ app/styles/themes/ocean.ts
@@ -1,0 +1,22 @@
+import { ThemeType } from "../../types/themes";
+
+export const ocean: ThemeType = {
+  name: "ocean",
+  background: "#dceff5",
+  backgroundSecondary: "#e8f6fa",
+  backgroundGradient: "linear-gradient(145deg, #eaf9fc, #c4e3ea)",
+  backgroundGradientTimer: "linear-gradient(145deg, #d5edf5, #a5cfd9)",
+  text: "#1d3d4f",
+  buttonText: "#13303f",
+  buttonTextDisabled: "#7a9da9",
+  buttonBackgroundActive: "#dceff5",
+  inputInner: "#e1f2f8",
+  highlight: "#007a99",
+  boxShadowInsetSoft: "inset 6px 6px 13px #bcd3d9, inset -6px -6px 13px #e7f5f7ff",
+  boxShadowOuter: "10px 10px 20px #a1b5bb, -10px -10px 22px #f0fdff",
+  boxShadowOuterStrong: "6px 6px 22px #99b1b7, -6px -6px 22px #e1f7fc",
+  boxShadowTimer: "inset 5px 5px 10px #b0cfd6, inset -5px -5px 10px #e7f5f7ff",
+  boxShadowTimerStrong: "15px 15px 30px #a2c2c9, -5px -5px 30px #d5f5fb",
+  boxShadowInputDisabled: "inset 1px 1px 5px #a2bcc1, inset -5px -5px 10px #e7f5f7ff",
+  boxShadowInput: "inset 5px 5px 7px #bedbe2, inset -5px -5px 7px #e7f5f7ff",
+};
