+++ app/styles/themes/peach.ts
@@ -1,0 +1,22 @@
+import { ThemeType } from "../../types/themes";
+
+export const peach: ThemeType = {
+  name: "peach",
+  background: "#f9ece6",
+  backgroundSecondary: "#faefebcc",
+  backgroundGradient: "linear-gradient(145deg, #fef5f1, #f1dcd4)",
+  backgroundGradientTimer: "linear-gradient(145deg, #f7e8e1, #e5c9be)",
+  text: "#2d1c17",
+  buttonText: "#3e2b26",
+  buttonTextDisabled: "#9e7e75f2",
+  buttonBackgroundActive: "#f9ece6",
+  inputInner: "#fdf5f1f2",
+  highlight: "#d17659",
+  boxShadowInsetSoft: "inset 6px 6px 13px #e5d6d0, inset -6px -6px 13px #fff9f6",
+  boxShadowOuter: "10px 10px 20px #d4c2bb, -10px -10px 22px #fff6f2",
+  boxShadowOuterStrong: "6px 6px 22px #d4c2bb, -6px -6px 22px #fff6f2",
+  boxShadowTimer: "inset 5px 5px 10px #decac3, inset -5px -5px 10px #f2e4df",
+  boxShadowTimerStrong: "15px 15px 30px #cbaea7, -5px -5px 30px #f2e4df",
+  boxShadowInputDisabled: "inset 1px 1px 5px #decac3, inset -5px -5px 10px #f2e4df",
+  boxShadowInput: "inset 5px 5px 7px #e1d0cb, inset -5px -5px 7px #f7efeb",
+};
