+++ app/styles/themes/primary.ts
@@ -1,0 +1,22 @@
+import { ThemeType } from "../../types/themes";
+
+export const primary: ThemeType = {
+  name: "primary",
+  background: "#e1dfdf",
+  backgroundSecondary: "#edececaa",
+  backgroundGradient: "linear-gradient(145deg, #f8f8f8, #ededed)",
+  backgroundGradientTimer: "linear-gradient(145deg, #f1efef, #cbc9c9)",
+  text: "#333333",
+  buttonText: "#252525",
+  buttonTextDisabled: "#878787f2",
+  buttonBackgroundActive: "#e1dfdf",
+  inputInner: "#eae7e7f2",
+  highlight: "#b62c0a",
+  boxShadowInsetSoft: "inset 6px 6px 13px #c8c6c6, inset -6px -6px 13px #faf8f8",
+  boxShadowOuter: "10px 10px 20px #989797, -10px -10px 22px #eeeded",
+  boxShadowOuterStrong: "6px 6px 22px #989797, -6px -6px 22px #eeeded",
+  boxShadowTimer: "inset 5px 5px 10px #bebebe, inset -5px -5px 10px #ffffff",
+  boxShadowTimerStrong: "15px 15px 30px #b6b6b6, -5px -5px 30px #e4e4e4",
+  boxShadowInputDisabled: "inset 1px 1px 5px #bebebe, inset -5px -5px 10px #eae6e6ff",
+  boxShadowInput: "inset 5px 5px 7px #c7c7c7, inset -5px -5px 7px #e0dadaff",
+};
