+++ app/styles/themes/sand.ts
@@ -1,0 +1,22 @@
+import { ThemeType } from "../../types/themes";
+
+export const sand: ThemeType = {
+  name: "sand",
+  background: "#f3ece6",
+  backgroundSecondary: "#f9f4f0",
+  backgroundGradient: "linear-gradient(145deg, #f7f1ed, #e9ded7)",
+  backgroundGradientTimer: "linear-gradient(145deg, #f4e7df, #d9c7bd)",
+  text: "#4d2c23",
+  buttonText: "#3e241d",
+  buttonTextDisabled: "#a18a82",
+  buttonBackgroundActive: "#f3ece6",
+  inputInner: "#f7f0eb",
+  highlight: "#bf5a3c",
+  boxShadowInsetSoft: "inset 6px 6px 13px #d6c8c1, inset -6px -6px 13px #fff9f4",
+  boxShadowOuter: "10px 10px 20px #b9aca6, -10px -10px 22px #fff6f1",
+  boxShadowOuterStrong: "6px 6px 22px #b09f97, -6px -6px 22px #f8eee9",
+  boxShadowTimer: "inset 5px 5px 10px #cab8b0, inset -5px -5px 10px #fffaf6",
+  boxShadowTimerStrong: "15px 15px 30px #c0aca5, -5px -5px 30px #f2e4dc",
+  boxShadowInputDisabled: "inset 1px 1px 5px #d4c4bc, inset -5px -5px 10px #fff9f5",
+  boxShadowInput: "inset 5px 5px 7px #e1d3cc, inset -5px -5px 7px #f8eee9",
+};
