+++ app/styles/themes/sunny.ts
@@ -1,0 +1,22 @@
+import { ThemeType } from "../../types/themes";
+
+export const sunny: ThemeType = {
+  name: "sunny",
+  background: "#fff4e5",
+  backgroundSecondary: "#fef2e0cc",
+  backgroundGradient: "linear-gradient(145deg, #fffaf3, #fbe5cf)",
+  backgroundGradientTimer: "linear-gradient(145deg, #fce7d5, #ecdfd4ff)",
+  text: "#3e2c1e",
+  buttonText: "#4a3625",
+  buttonTextDisabled: "#a58a74f2",
+  buttonBackgroundActive: "#fff4e5",
+  inputInner: "#fff6ebf2",
+  highlight: "#ff8f3c",
+  boxShadowInsetSoft: "inset 6px 6px 13px #e0cbb8, inset -6px -6px 13px #d6d6d6ff",
+  boxShadowOuter: "10px 10px 20px #d4c3b2, -10px -10px 22px #fffaf4",
+  boxShadowOuterStrong: "6px 6px 22px #d4c3b2, -6px -6px 22px #fffaf4",
+  boxShadowTimer: "inset 5px 5px 10px #e2cfbc, inset -5px -5px 10px #d6d6d6ff",
+  boxShadowTimerStrong: "15px 15px 30px #c9b29a, -5px -5px 30px #fef0e3",
+  boxShadowInputDisabled: "inset 1px 1px 5px #e2cfbc, inset -5px -5px 10px #d6d6d6ff",
+  boxShadowInput: "inset 5px 5px 7px #eedcca, inset -5px -5px 7px #d6d6d6ff",
+};
